backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
google-genai==1.5.0
python-dotenv==1.0.1
python-multipart==0.0.12
